refactor(backend): postRelations를 별도 파일로 분리
코드 리뷰 반영. shared 위치인 post.entity.ts가 RecruitmentDetail,
TipsDetail, Scrap, Comment 등 구체 도메인을 import하면서 의존성
방향이 역전돼 있었음. Feed 도메인이 추가되면 shared가 모든 형제
모듈을 알아야 하는 강한 결합 문제로 이어짐.

postRelations만 post/post-relations.ts로 분리해서 Post entity는
순수 테이블 정의(+ Member FK)만 갖도록 정리. relations metadata는
cross-cutting이라 구체 도메인 import OK.
