+++ apps/backend/src/common/db/schema/index.ts
@@ -5,6 +5,7 @@ export * from '@/member/domain';
 export * from '@/report/domain';
 export * from '@/post/shared/domain';
 export * from '@/post/domain';
+export * from '@/post/post-relations';
 export * from '@/comment/domain';
 export * from '@/participation/domain';
 export * from '@/scrap/domain';
