+++ apps/backend/src/post/shared/domain/post.entity.ts
@@ -1,12 +1,6 @@
 import { integer, pgTable, text } from 'drizzle-orm/pg-core';
 import { baseColumns } from '@/common';
 import { Member } from '@/member/domain';
-import { relations } from 'drizzle-orm';
-import { TipsDetail } from '@/post/domain/tips-detail.entity';
-import { RecruitmentDetail } from '@/post/domain/recruitment-detail.entity';
-import { PostImage } from '@/post/shared/domain/post-image.entity';
-import { Scrap } from '@/scrap/domain';
-import { Comment } from '@/comment/domain';
 import { postTypePgEnum, postStatusPgEnum } from '@/common';
 
 export { postStatusPgEnum, postTypePgEnum };
@@ -20,15 +14,3 @@ export const Post = pgTable('post', {
     .notNull()
     .references(() => Member.id),
 });
-
-export const postRelations = relations(Post, ({ one, many }) => ({
-  author: one(Member, {
-    fields: [Post.authorId],
-    references: [Member.id],
-  }),
-  recruitmentDetail: one(RecruitmentDetail),
-  tipsDetail: one(TipsDetail),
-  postImage: many(PostImage),
-  scrap: many(Scrap),
-  comment: many(Comment),
-}));
